[REFACTOR] #119 팔로우, 스크랩, 공감 api 예외처리 추가

Closes #119

See merge request s08-webmobile2-sub2/S08P12A709!131

Former-commit-id: 02a4115c82854fe12500ce14444cf1144e12edde

+++ backend/mindderStart/src/main/java/com/ssafy/mindder/my/controller/MyController.java
@@ -165,7 +165,9 @@ public class MyController {
 		logger.debug("followAdd - 호출 : " + targetUserIdx);
 		try {
 			int userIdx = jwtService.getUserIdx(accessToken);
-			
+			if (myService.findFollow(userIdx, targetUserIdx) != null) {
+				return ApiResponse.error(ErrorCode.VALIDATION_FOLLOW_EXCEPTION);
+			}
 			myService.addMyFollow(userIdx, targetUserIdx);
 			return ApiResponse.success(SuccessCode.CREATE_MY_FOLLOW);
 		} catch (Exception e) {
@@ -183,7 +185,9 @@ public class MyController {
 		logger.debug("followRemove - 호출 : " + targetUserIdx);
 		try {
 			int userIdx = jwtService.getUserIdx(accessToken);
-			
+			if (myService.findFollow(userIdx, targetUserIdx) == null) {
+				return ApiResponse.error(ErrorCode.NOT_FOUND_FOLLOW_EXCEPTION);
+			}
 			myService.removeMyFollow(userIdx, targetUserIdx);
 			return ApiResponse.success(SuccessCode.DELETE_MY_FOLLOW);
 		} catch (Exception e) {
